Fix trivy scan failures w/ otel upgrades (#5906)
* engine: upgrade otel deps to address trivy scan result

https://github.com/dagger/dagger/security/dependabot/381

Signed-off-by: Erik Sipsma <erik@dagger.io>

* engine: log+ignore otel tracer errors

If there's mismatches between dependency versions on otel libs between
us and buildkit (and its various deps), then we can get schema merge
errors.

While sad and worth notifying the user about via a log statement, these
are not actually fatal, so it's not worth exiting over.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* core: ignore lint failure about deprecated jaeger exporter import

Signed-off-by: Erik Sipsma <erik@dagger.io>

* assert on more specific message in internal pipeline test

This is a bit silly. The previous update to log otel failures caused the
word "merge" to appear in the output (it's an error about merging otel
schemas), which in turn caused our test assertion to fail because it's
checking that no "merge" (-op) vertexes show up in the output by
default.

There's not really any great way to make the test more robust than I can
see other than just asserting on a more specific string, "merge (". It's
important to have this test coverage though, so this seems like the best
solution.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* ci: use buildctl binary from upstream image

See comment in `buildctlBin` func in `internal/mage/util/engine.go`
for reasoning.

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>