core: fix huge module/func IDs by removing Function.call (#5947)
* core: fix huge module/func IDs by removing Function.call

We are hitting some problems with IDs growing to enormous size and
exceeding gRPC messages limits in relatively normal cases (i.e. 13
functions defined in a module).

I didn't fully track down exactly what was happening, but suspected that
it was related to the annoyance we were dealing with internally to store
ModuleID on the Function object.

That annoyance only existed in order to support the Function.call
"dynamic" API. That API isn't actually used anywhere right now, so I just
tried removing that API call and the associated ModuleID handling in
Function and sure enough the problem w/ huge IDs went away.

This feels like the best solution for now. If a request for the
Function.call API arises in the future it can be re-added but rather
than adding it to the Function object, it can be put on the Module
object (and just require object+func name in addition to existing args),
which should avoid the problem we're hitting now.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* regen sdks

Signed-off-by: Erik Sipsma <erik@dagger.io>

* remove unused Function.ParentName field

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>