core: namespace module objects and function extensions (#5932)
* core: namespace module objects and function extensions

This change results in names being prefixed by module name for:
1. Any objects defined by the module
2. Any functions defined by the module that are extending objects from
   other pre-existing modules (or core)

This reduces the chance of conflicts when depending on 2+ modules that
happen to use overlapping names.

It's implemented so that SDKs don't need any awareness of it; when a
runtime is called names are "de-namespaced" as needed so the runtime
receives names the same as what it originally defined.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* update experimental zenith docs w/ namespaced function

Signed-off-by: Erik Sipsma <erik@dagger.io>

* ci: remove conflicting cache mounts from rust test

We run CI on dagger runners now, which allows for much greater sharing
of cache mounts between different runs in parallel.

I saw some fairly consistent failures in the Rust SDK tests which seemed
to be related to missing files from cache mounts.

The cache mounts in question did not seem safe to be shared across
pipelines since they relied on the existence of a single file (as
opposed to something like cargo cache which I understand to be more like
a content addressed store or sorts).

This probably wasn't safe before either since cache mounts can be pruned
at any moment, so it's not safe to assume their contents will be present
in later execs.

Fixed this by removing some problematic cache mounts that couldn't be
namespaced and by namespacing the cargo cache mount with rust version
(in order to prevent conflicts between different versions of rust
running in parallel on the same dagger runner).

Signed-off-by: Erik Sipsma <erik@dagger.io>

* fix use of ~ in rust cargo cache

Signed-off-by: Erik Sipsma <erik@dagger.io>

* ci: force gzip compression in exported ./hack/dev engine tar

Signed-off-by: Erik Sipsma <erik@dagger.io>

* ci: combine rust steps to remove assumptions about cache volume state

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>