python: rename modules package (#5962)
It used to be "dagger.server", then we changed to environments and it was renamed to "dagger.env", and finally, before the change to modules, this was renamed more generally to "dagger.ext" (for extensions).

Renaming again to "dagger.mod", and removing the previous concept of multiple types of entrypoints to just "functions".

Signed-off-by: Helder Correia <174525+helderco@users.noreply.github.com>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>