Add DAGGER_CLOUD_TOKEN client-side so that we get telemetry in Cloud (#6020)
* ci: Add DAGGER_CLOUD_TOKEN client-side

Without this, even though the Engine already has this env var, runs will
**NOT** appear in https://dagger.cloud

To test the above, I have not yet configured this secret in the
repository. I will do a follow-up commit after the secret is set. If the
above is correct, the run for the first commit will not show in
https://dagger.cloud

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* ci: Replace _EXPERIMENTAL_DAGGER_CLOUD_TOKEN with DAGGER_CLOUD_TOKEN

This is a legacy token that was replaced with DAGGER_CLOUD_TOKEN which
should be kept secret. FWIW: https://docs.dagger.io/cloud/572923/get-started

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>