engine: update to latest graphql-go version (#6056)
* engine: update to latest graphql-go version

The change there makes the return value of graphql introspection stable,
which is important since it's an input to the SDK Codegen+Runtime
functions.

Previously, we were re-running those SDK functions way too often, which
wasted tons of time, especially in modules with larger dependency DAGs.

One of Kyle's demo's took a minimum of 25s every run due to this; now
it's down to 7s with this fix in place.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* update golangci version to fix incorrect lint failure

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>