Bump otelgrpc to fix CVE issue (#6100)
* Bump otelgrpc to v0.46.0

This fix one of issues, founded in trivy scan job [1].

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>

* Run go mod tidy to clean up dependencies

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>

* Bump go.opentelemetry.io/otel/sdk to 1.20.0

Since bump otelgrpc, it's required to bump otel/sdk to 1.20.

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>

* engine: use our own dial-stdio binary rather than buildctl

The buildctl binary is currently required in the engine image since it
facilitates connhelpers connecting to the engine inside (by proxying the
stdio of an exec'd buildctl process to the unix socket). This is
inherited from buildkit and how Dagger has worked for a while.

The problem is we keep running into vulnerabilities in this binary that
take a while for upstream to fix (i.e. related to otel/grpc), which
leaves us to resort to various hacks to get a binary w/out
vulnerabilities in the engine image.

This change moves us off of upstream's buildctl binary and instead
replaces it with the absolute minimal features we need in
`./cmd/dialstdio`. There's none of the buildctl subcommands except
dial-stdio. This allows us to easily build it from our own go.mod (which
makes dependency upgrades easier) and also results in most problematic
sources of dependencies no longer being present.

The source code for the dial-stdio functionality is mostly a fork of
upstream, except all the other adjacent code we don't need has been
stripped out. I also switched to use cobra to be consistent w/ our other
binaries and avoid a separate dependency on urfave/cli.

Attribution on the original source code is given w/ a NOTICE and
BUILDKIT_AUTHORS file, same as we did w/ `./cmd/engine`.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* ci: invalidate apk install of engine image once per day.

We currently are in a situation where the engine image is being built
and installing apk packages that are vulnerable. Alpine released newer
versions of the packages a few days ago, but we are still hitting cache
for the "apk add" execs from before then and thus aren't picking them
up.

The real fix to this is to switch CI to use Zenith and use one of the
various apk/apko/wolfi modules that have been written which avoid this
problem, e.g.
https://daggerverse.dev/mod/github.com/sipsma/daggerverse/apk@d709dd631e176025ea042dc03fdc90197f4b12ba

However, in the meantime this change just wraps the apk add w/ an env
var that changes once per day (current unix time / 1 day in seconds),
which will hold us over until then.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* engine: ignore deprecation lints about otelgrpc interceptors.

It's not immediately clear how to replace the interceptors with the
suggested "NewServerHandler" while still working w/ grpc_middleware.

It's possible that's already do-able and I just haven't seen exactly how
yet; I didn't invest a ton of time into this since we aren't really
using otel yet. Upstream buildkit will also need to deal with this once
they upgrade so we can see what they end up doing too.

Signed-off-by: Erik Sipsma <erik@dagger.io>

* Use time.Truncate to simplify apk cache expiry

Co-authored-by: Alex Suraci <suraci.alex@gmail.com>
Signed-off-by: Erik Sipsma <erik@sipsma.dev>

---------

Signed-off-by: Thanabodee Charoenpiriyakij <wingyminus@gmail.com>
Signed-off-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Erik Sipsma <erik@sipsma.dev>
Co-authored-by: Erik Sipsma <erik@dagger.io>
Co-authored-by: Erik Sipsma <erik@sipsma.dev>
Co-authored-by: Alex Suraci <suraci.alex@gmail.com>