Zenith (Go SDK): allow omitting JSON types in structs (#6057)
* fix: capture field value at start of loop

Somehow this wasn't causing massive bugs? This meant we were getting the
wrong name/type in the resolver.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* feat: allow OriginalName for fields

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: go codegen should request the json field name

...if it's present. The server now correctly sends requests to the
module that match what it asks for (regardless of what casing the user
wanted). So we now need to ask for the right thing.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* test: add test for structs with json combinations

Signed-off-by: Justin Chadwell <me@jedevc.com>

* test: fixup test to detect id instead of ID

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: allow nested structs from module functions

Previously this would just fail (but would allow nesting lists of
objects!). This was because the condition we were checking didn't take
into account non-idable object types (which is what's returned from
modules).

Also, this prevents the previously weird case of allowing a module to
declare it returns a directory and then return a *different* core type,
like a container, which would actually not error immediately.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* fix: allow returning lists of core-idable types

This also applies to fields that are lists of core-idable types.

Essentially, this translates to returning a list of ids, which we need
to lift to a list of objects. The easiest way to do this is by having a
little helper func.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* test: add new test cases suggested in code review

Signed-off-by: Justin Chadwell <me@jedevc.com>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>