feat: make module types id-able  (#6102)
* feat: make module types id-able

This feature allows all module declared objects to be converted to a
respective ID-type using the "id" field, and converted from an ID-type
using the "loadFromID" helper. This effectively makes these types
fully idable.

IDs are essentially the contents of the module type's data,
JSON-encoded, and then base64-encoded, similar (but not identical) to
standard IDs. At the moment, they don't share exactly the same codepaths
to normal modules, but this could be adapted to be so in the future. We
also include a note of *what type* generated this ID - this is so we can
prevent type confusion.

Essentially, this enables a number of previously impossible use cases:

1. Returning a custom object from another module from a function

   - This case is the simplest - we just extend our conversion that
     unpacks all returned IDs from functions into their underlying types
     (so while a module returns an XID, it can be queried as if it was X).

2. Accepting a custom object as a parameter to a function

   - Because graphql doesn't support creating inline objects, we get
     around this in the core api by instead passing IDs everywhere. All
     the code-gen is setup to expect this, but module types weren't
     IDable, so we failed during schema-checking.

     Additionally, we need to transparently convert the ID from the
     graphql request into the underlying storage data for the module.

3. Accepting a custom object from another module as a parameter to a
   function

   - This is the same case as above, but also involves ensuring that we
     stitch in the type declarations from the module into our own schema
     so that we have the right type.

     However, here, we **don't** decode the ID into its underlying
     data, it's a foreign type, so the module should only see the ID.

Signed-off-by: Justin Chadwell <me@jedevc.com>

* disallow use of external module types in function+field signatures

While we now give all module objects ID-ability and enable a module to
accept its own objects as function inputs, we are not yet going to allow
modules to use objects from other (non-core) modules as function args,
function returns or object fields.

This is because allowing that results in a caller who depends on a given
module A to now not only have access to module A's types but also the
types of transitive dependencies. This in turn opens the doors to
dependency hell; if two versions of a given module show up multiple
times in the full dependency DAG we need some fairly complex dependency
resolution logic around which of those versions is exposed to the
caller.

That problem is famously difficult and while we may need to deal with it
someday, it's preferable to avoid it for as long as possible.

The idea is to instead add support for interfaces in follow ups, which
should unblock many of the use cases that call for interoperability
between different module types.

The main changes in this commit are to detect when a user tries to do
this and give them a descriptive error message in that case. Other
related changes are:
* an additional integ test case for the situation where there's
  conflicting versions of a give module in the transitive dependency
  DAG, which is expected to continue working.
* tracking of which module a given schema type was loaded from, which
  allows us to distinguish between use of a type from core (which is
  still allowed) and a type from a dependency module.
* removal of "Dependencies" from the ExecutableSchema interface we use
  internally, which overlapped with some of the other changes here and
  was completely unused (leftover from ancient days during the Cloak
  project).

Signed-off-by: Erik Sipsma <erik@dagger.io>

* check expected module type name direct in decode

Signed-off-by: Erik Sipsma <erik@dagger.io>

---------

Signed-off-by: Justin Chadwell <me@jedevc.com>
Signed-off-by: Erik Sipsma <erik@dagger.io>
Co-authored-by: Erik Sipsma <erik@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>