website: Fix a bunch of Dependabot security alerts (#6133)
* Force update @babel/traverse dependency in website

It had to be manually deleted from the yarn.lock file so that the latest
one could be installed. Followed the instructions from
https://github.com/dagger/dagger/security/dependabot/380 more closely.

    ❯ yarn list --pattern @babel/traverse
    yarn list v1.22.19
    └─ @babel/traverse@7.23.4
    ✨  Done in 0.57s.

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* Update all website deps at once

Meant to address a bunch of CVEs.

Removing the yarn.lock and running `yarn install` again was the most
efficient way of getting this done. Everything seems to be working fine
locally, let's push this and see what happens on the preview
environment.

As a follow-up, we should upgrade to Docusaurus v3 so that we solve a
few more of the remaining CVEs:
https://docusaurus.io/docs/migration/v3

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>

* website: Fix missing double quotes
This was breaking the build

Signed-off-by: jffarge <jf@dagger.io>

---------

Signed-off-by: Gerhard Lazu <gerhard@dagger.io>
Signed-off-by: jffarge <jf@dagger.io>
Co-authored-by: jffarge <jf@dagger.io>
Signed-off-by: Christian Schlatter <schlatter@puzzle.ch>